requirements.txt
# Ticket triage ingestion module
pydantic>=2.0,<3
python-dateutil>=2.8,<3
fastapi>=0.109,<1
uvicorn[standard]>=0.27,<1
